+++ changelog.md
@@ -2,6 +2,36 @@
 
 ---
 
+## v0.0.8
+
+* Fresh upload of app code due to install trigger issues.
+
+---
+
+## v0.0.7
+
+* Troubleshooting triggers due to app install error.
+
+---
+
+## v0.0.6
+
+* Troubleshooting triggers due to app install error.
+
+## v0.0.5
+
+* Troubleshooting triggers due to app install error.
+
+## v0.0.4
+
+* Updated imports.
+
+## v0.0.3
+
+* Added expanded console logs to pinpoint trigger issues.
+
+---
+
 ## v0.0.2
 
 * Resolved issue with install and upgrade triggers.